% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{add_legend}
\alias{add_legend}
\title{Add legend of pseudotime value}
\usage{
add_legend(pstime)
}
\arguments{
\item{pstime}{a numeric vector of pseudotime value}
}
\description{
Add continuous legend to pseudotime value plot
}
\examples{
\dontrun{add_legend(c(32.3,45.3,23.5,1.2,55.6))}
}
