% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood_meth_data}
\alias{blood_meth_data}
\title{Methylation matrix of 6251 blood samples after preprocess}
\format{
A data frame with 9447 rows and 6251 variables:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
blood_meth_data
}
\description{
A dataset containing the methylation levels of 6251 samples at 9447 CpG sites
}
\keyword{datasets}
