% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_meta_data}
\alias{brain_meta_data}
\title{Meta data of 675 brain methylation array samples}
\format{
A data frame with 675 rows and 5 variables:
\describe{
\item{sample_labels}{label of sample}
\item{chrono_age}{chronological age of sample}
\item{epigen_age}{epigenetic state of sample}
\item{exp_labels}{experiment or group label of sample}
\item{train_test_labels}{label of whether sample is in train/val/test set}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
brain_meta_data
}
\description{
A dataset containing the meta data of 675 methylation array samples
}
\keyword{datasets}
