% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_meth_data}
\alias{brain_meth_data}
\title{Methylation matrix of 675 brain samples after preprocess}
\format{
A data frame with 8510 rows and 675 variables:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
brain_meth_data
}
\description{
A dataset containing the methylation levels of 675 samples at 8510 CpG sites
}
\keyword{datasets}
