% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{check_valid_data}
\alias{check_valid_data}
\title{Check valid input}
\usage{
check_valid_data(meth_data, meta_data)
}
\arguments{
\item{meth_data}{a data matrix
\describe{ Each column contains methylation data of each sample.
Each row contains methylation data at each site.
}}

\item{meta_data}{a data frame
\describe{Each row contains meta data of each sample (NA if not applicable).
Each column contains meta data in the following order.
\itemize{
\item chrono_age : numeric chronological age
\item epigen_age : numeric epigenetic age
\item exp_labels : experiment label
\item train_test_labels : label of whether the sample is in train/test/validation set
}}}
}
\value{
a list of data frame such that
list("meth_data" = meth_data, "meta_data" = meta_data)
}
\description{
Check dimension and data type of input data frames if valid
}
\examples{
\dontrun{check_valid_data(meth_data, meta_data)}

}
