% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainTestSplit.R
\name{exp_train_val_test_split}
\alias{exp_train_val_test_split}
\title{Train-validation-test split by experiment label}
\usage{
exp_train_val_test_split(data_list, train_exp, test_exp, val_exp)
}
\arguments{
\item{data_list}{a list of methylation and meta data data frame respectively}

\item{train_exp}{a character vector of experiments in train set}

\item{test_exp}{a character vector of experiments in test set}

\item{val_exp}{(optional) a character vector of experiments in val set}
}
\value{
a list of 4 or 6
\describe{
\itemize{
\item train_meth : training data
\item test_meth: test data
\item val_meth (optional) : validation data
\item train_meta : training meta data
\item test_meta : test meta data
\item val_meta (optional): validation meta data
}
}
}
\description{
Split data into train-val-test set by given experiment label
}
\examples{
\dontrun{exp_train_val_test_split(list("meth_data" = meth_data, "meta_data" = meta_data),
                         train_exp = c('GSE482504', 'GSE235456'),
                         test_exp = c('GSE621336','GSE125673'),
                         val_exp = c('GSE34354'))}

}
