% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitPseudotime.R
\name{fit_pseudotime}
\alias{fit_pseudotime}
\title{Fit Pseudotime Model}
\usage{
fit_pseudotime(
  train_meth,
  train_meta,
  clust_num,
  tissue_type = "",
  model_label = "",
  choose_young_clust = FALSE,
  plot = TRUE,
  plot_trend = FALSE,
  start_plot_fit = c(a0 = 9.79, b1 = 2, c1 = 0.025, c2 = 0.412)
)
}
\arguments{
\item{train_meth}{a data frame of training methylation data}

\item{train_meta}{a data frame of meta data for the training samples}

\item{clust_num}{(optional) a numeric value of cluster number}

\item{tissue_type}{(optional) a string of tissue type for plot title}

\item{model_label}{(optional) a string of model label for plot file name}

\item{choose_young_clust}{(optional) a boolean whether to use cluster with
youngest mean age as starting cluster of the trajectory}

\item{plot}{(optional) a boolean of whether to plot pseudotime trajectory
and pseudotime vs chronological age and save plots to disk}

\item{plot_trend}{(optional) a boolean of whether to fit and plot
nls best fit sum of two exponential trend line for pseudotime vs age}

\item{start_plot_fit}{(optional) a named vector of starting value for nls
fitting sum of two exponential trend to pseudotime vs age}
}
\value{
a list of 2
\describe{
\itemize{
\item "reduced_dim_model" prcomp model object used in model training
\item "pstime_model" trained slingshot model object
}
}
}
\description{
Fit slingshot pseudotime model on training data
}
\examples{
\dontrun{
fit_pseudotime(train_meth, train_meta, tissue_type = "Blood",
               model_label = "train_ver1", plot = TRUE)
}
}
