% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitPseudotime.R
\name{get_single_pseudotime}
\alias{get_single_pseudotime}
\title{Get single pseudotime value}
\usage{
get_single_pseudotime(pseudotime_obj)
}
\arguments{
\item{pseudotime_obj}{trained slingshot model object}
}
\value{
an array of pseudotime value for each data points
}
\description{
Get single pseudotime value for each data points.
If there are more than 1 linages, the pseudotime is calculated
as a weighted average across all linage.
}
\examples{
\dontrun{get_single_pseudotime(slingshot_model)}
}
