% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainTestSplit.R
\name{labels_train_test_val_split}
\alias{labels_train_test_val_split}
\title{Train-validation-test split by train-test label}
\usage{
labels_train_test_val_split(data_list, train_label, test_label, val_label)
}
\arguments{
\item{data_list}{a list of methylation and meta data data frame respectively}

\item{train_label}{a string keyword of train set label}

\item{test_label}{a string keyword of test set label}

\item{val_label}{(optional) a string keyword of val set label}
}
\value{
a list of 4 or 6
\describe{
\itemize{
\item train_meth : training data
\item test_meth: test data
\item val_meth (optional) : validation data
\item train_meta : training meta data
\item test_meta : test meta data
\item val_meta (optional): validation meta data
}
}
}
\description{
Split data into train-val-test set by train-test label
}
\examples{
\dontrun{labels_train_test_val_split(list("meth_data" = meth_data, "meta_data" = meta_data),
                         train_label = "train",
                         test_label = "test",
                         val_label = "val")

labels_train_test_val_split(list("meth_data" = meth_data, "meta_data" = meta_data),
                         train_label = "train",
                         test_label = "test")}

}
