% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitPseudotime.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(y, x)
}
\arguments{
\item{y}{a numeric vector of predicted data}

\item{x}{a numeric vector of observed data}
}
\value{
Mean absolute error between x and y
}
\description{
Calculate mean absolute error
}
\examples{
\dontrun{mae(c(1,2,3),c(4,5,6))}
}
