% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_pseudotime_cluster}
\alias{plot_pseudotime_cluster}
\title{Plot pseudotime trajectory over cluster}
\usage{
plot_pseudotime_cluster(
  pstime_model,
  reduced_dim,
  clusters,
  tissue_type = "",
  model_label = ""
)
}
\arguments{
\item{pstime_model}{a slingshot pseudotime object}

\item{reduced_dim}{a dataframe of training data in low dimensional space}

\item{clusters}{a vector of cluster labels of reduced_dim}

\item{tissue_type}{(optional) a string of tissue type for plot title}

\item{model_label}{(optional) a string of label for plot file name}
}
\description{
Plot pseudotime trajectory over clustered train data points
}
\examples{
\dontrun{plot_pseudotime_value(slingshot_model,
                      reduced_dim,
                      clusters,
                      tissue_type = "Blood",
                      model_label = "train_v1")}
}
