% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_pseudotime_value}
\alias{plot_pseudotime_value}
\title{Plot pseudotime trajectory with pseudotime value}
\usage{
plot_pseudotime_value(
  pstime_model,
  reduced_dim,
  tissue_type = "",
  model_label = ""
)
}
\arguments{
\item{pstime_model}{a slingshot pseudotime object}

\item{reduced_dim}{a dataframe of training data in low dimensional space}

\item{tissue_type}{(optional) a string of tissue type for plot title}

\item{model_label}{(optional) a string of label for plot file name}
}
\description{
Plot pseudotime trajectory with pseudotime value of each data point
}
\examples{
\dontrun{plot_pseudotime_value(slingshot_model,
                      reduced_dim,
                      tissue_type = "Blood",
                      model_label = "train_v1")}
}
