% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot data points}
\usage{
plot_scatter(
  x,
  y,
  title,
  group,
  trend_model,
  func_form = c("exp", "quad", "log", "sqrt", "doubleExp"),
  tissue_type = "",
  model_label = "",
  xlab = "x",
  ylab = "y",
  grouplab = "group",
  filename = "scatter_plot",
  width = 2100,
  height = 1500
)
}
\arguments{
\item{x}{a numeric vector of x value}

\item{y}{a numeric vector of y value}

\item{title}{a string of plot title}

\item{group}{(optional) a vector of group labels of data}

\item{trend_model}{(optional) fitted (lm/nls) model of the trend line}

\item{func_form}{(optional) functional form of the trend line}

\item{tissue_type}{(optional) a string of tissue type for plot title}

\item{model_label}{(optional) a string of label for plot file name}

\item{xlab}{(optional) a string of x label}

\item{ylab}{(optional) a string of y label}

\item{grouplab}{(optional) a string of label for group labels}

\item{filename}{(optional) a string of plot file name}

\item{width}{(optional) a numeric value of width of the plot in pixel}

\item{height}{(optional) a numeric value of height of the plot in pixel}
}
\description{
Make scatter plot of data points
}
\examples{
\dontrun{plot_scatter(x = chrono_age,
             y = pseudotime_value,
             title = "Pseudotime VS Chronological Age",
             group = exp_labels,
             tissue_type = "Blood",
             model_label = "testv.1",
             xlab = "Chronological Age (yrs)", ylab = "Pseudotime",
             grouplab = "experiment",
             filename = "epigenage",
             width = 2600,
             height = 1350)}
}
