% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictPseudotime.R
\name{predict_pseudotime}
\alias{predict_pseudotime}
\title{Predict Pseudotime value}
\usage{
predict_pseudotime(
  test_meth,
  test_meta,
  model_list,
  tissue_type = "",
  model_label = "",
  plot = TRUE,
  plot_trend = FALSE,
  start_plot_fit = c(a0 = 9.79, b1 = 2, c1 = 0.025, c2 = 0.412)
)
}
\arguments{
\item{test_meth}{a data frame of methylation data}

\item{test_meta}{a data frame of meta data}

\item{model_list}{a list of prcomp model object used in model training and
a slingshot object returned from CV_pseudotime or fit_pseudotime}

\item{tissue_type}{(optional) a string of tissue type for plot title}

\item{model_label}{(optional) a string of model label for plot file name}

\item{plot}{(optional) a boolean of whether to plot pseudotime trajectory
and pseudotime vs chronological age and save plots to disk}

\item{plot_trend}{(optional) a boolean of whether to fit and plot
nls best fit sum of two exponential trend line for pseudotime vs age}

\item{start_plot_fit}{(optional) a named vector of starting value for nls
fitting sum of two exponential trend to pseudotime vs age}
}
\value{
a numeric vector of predicted pseudotime value
}
\description{
Predict pseudotime value of a dataset given a pseudotime model
}
\examples{
\dontrun{predict_pseudotime(test_meth, test_meta,
                   list("reduced_dim_model" = prcomp_model,
                        "pstime_model" = slingshot_model),
                   tissue_type = "Blood",
                   model_label = "test_ver1",
                   plot = TRUE)}
}
