% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{print_sample_num_by_exp}
\alias{print_sample_num_by_exp}
\title{print sample number in an experiment}
\usage{
print_sample_num_by_exp(meta_data, exp_label)
}
\arguments{
\item{meta_data}{a data frame of meta data}

\item{exp_label}{string of experiment label}
}
\description{
print sample number of given experiment label
}
\examples{
\dontrun{print_sample_num_by_train_test(meta_data,'GSE42861')}
}
