% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{print_sample_num_by_train_test}
\alias{print_sample_num_by_train_test}
\title{print sample number by experiment in train/test set}
\usage{
print_sample_num_by_train_test(meta_data, train_test_label)
}
\arguments{
\item{meta_data}{a data frame of meta data}

\item{train_test_label}{a string for train, val, or test label}
}
\description{
print sample number in each experiment of the
given train-test label
}
\examples{
\dontrun{print_sample_num_by_train_test(meta_data,"test")
print_sample_num_by_train_test(meta_data,"train")}
}
