% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitPseudotime.R
\name{r2}
\alias{r2}
\title{R squared}
\usage{
r2(preds, actual)
}
\arguments{
\item{preds}{a numeric vector of predicted data}

\item{actual}{a numeric vector of observed data}
}
\value{
R squared value between preds and actual
}
\description{
Calculate R squared value
}
\examples{
\dontrun{r2(c(1,2,3),c(4,5,6))}
}
