% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainTestSplit.R
\name{ratio_train_val_test_split}
\alias{ratio_train_val_test_split}
\title{Stratified train-validation-test split by ratio}
\usage{
ratio_train_val_test_split(
  data_list,
  train_size,
  test_size,
  val_size = 0,
  age_bin_num = 20
)
}
\arguments{
\item{data_list}{a list of methylation and meta data data frame respectively}

\item{train_size}{a numeric value of ratio to split into train set}

\item{test_size}{a numeric value of ratio to split into the test set}

\item{val_size}{(optional) a numeric value of ratio to split into val set}

\item{age_bin_num}{number of binnings of chronological age for stratification}
}
\value{
a list of 4 or 6
\describe{
\itemize{
\item train_meth : training data
\item test_meth: test data
\item val_meth (optional) : validation data
\item train_meta : training meta data
\item test_meta : test meta data
\item val_meta (optional): validation meta data
}
}
}
\description{
Split data into train-val-test set, stratified by age bin.
}
\examples{
\dontrun{ratio_train_val_test_split(list("meth_data" = meth_data, "meta_data" = meta_data),
                           train_size = 0.2,
                           test_size = 0.6, val_size = 0.1,
                           age_bin_num = 25)
}
}
