% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{remove_bad_exp}
\alias{remove_bad_exp}
\title{Remove unwanted experiment}
\usage{
remove_bad_exp(data_list, bad_exps)
}
\arguments{
\item{data_list}{a list of methylation and meta data data frame respectively}

\item{bad_exps}{a character vector of experiment labels}
}
\value{
\describe{
a list of methylation and meta data data frames without unwanted experiment
such that list("meth_data" = meth_data, "meta_data" = meta_data)
}
}
\description{
Remove unwanted experiment given a vector of experiment labels
}
\examples{
\dontrun{remove_bad_exp(list("meth_data" = meth_data, "meta_data" = meta_data),
               c('GSE45363','GSE34564'))}
}
