% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitPseudotime.R
\name{rmse}
\alias{rmse}
\title{Root mean square error}
\usage{
rmse(y, x)
}
\arguments{
\item{y}{a numeric vector of predicted data}

\item{x}{a numeric vector of observed data}
}
\value{
root mean square error between x and y
}
\description{
Calculate root mean square error
}
\examples{
\dontrun{rmse(c(1,2,3),c(4,5,6))}
}
