% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSite.R
\name{select_site_var}
\alias{select_site_var}
\title{Select site by variance}
\usage{
select_site_var(meth_data, meta_data, var_thresh = 0.001)
}
\arguments{
\item{meth_data}{a data frame of methylation data}

\item{meta_data}{a data frame of meta data}

\item{var_thresh}{a numeric value of variance threshold}
}
\value{
a vector of index that satisfy the variance threshold
}
\description{
Select site with larger Pearson's Correlation Coefficient
with chronological age than the given threshold
}
\examples{
\dontrun{select_site_var(meth_data, meta_data)
select_site_var(meth_data, meta_data, var_thresh = 0.002)}
}
